Remove `substrate-wasm-builder`
Usage removed with
https://github.com/serai-dex/serai/commit/b79125664894450693a28cb6d6e4358778b7968c.

We had a patch for it propagating polluted environment variables, a patch to
make `polkavm-linker` an optional dependency, enforcement of `wasm32v1-none`
and `wasm32v1-none` alone (not only for its stronger properties, yet because
detection of the toolchain failed on Alpine for whatever reason), and various
patches for its dependencies (`filetime`, `build-helper`) with an open issue
for another dependency, `walkdir`
(https://github.com/serai-dex/serai/issues/4).

While replacing it is non-trivial, so it can be explained as to _why_ such a
crate exists, the amount of cruft it has over the years can't be justified.
While we already made notable patches, we had an explicit patch wanted, and
likely more along the way.

This builder also took multiple minutes to compile, having over three hundred
dependencies, so its removal translates to real-world benefit to our CI,
testing, and developer experience.

Closes https://github.com/serai-dex/serai/issues/4.