Remove all extensions from `frame-system`
The remaining extensions were primarily about the weight (and size) of
transactions. Serai did not adopt the transaction extension framework, and
while the check weight functionality was kept as it would still have been
useful, it was re-implemented with
https://github.com/serai-dex/serai/commit/ff95c5834134e26f9569e77e7482058999cf0673.

With this commit to not using extensions, as integrating the one (set of)
extension(s) we wanted wasn't worth the effort, we can simplify to pruning the
entire folder of provided extensions.